Switching serialization quirks to uint64_t for consistency with other bitfields

+++ core.h
@@ -178,9 +178,9 @@ bool core_set_environment(retro_ctx_environ_info_t *info);
 
 bool core_serialize_size(retro_ctx_size_info_t *info);
 
-uint32_t core_serialization_quirks(void);
+uint64_t core_serialization_quirks(void);
 
-void core_set_serialization_quirks(uint32_t quirks);
+void core_set_serialization_quirks(uint64_t quirks);
 
 bool core_serialize(retro_ctx_serialize_info_t *info);
